updated some references and some minor updates.

+++ ali_dindin/models/res_config_settings.py
@@ -82,5 +82,6 @@ class ResConfigSettings(models.TransientModel):
         else:
             cron = self.env['ir.cron'].sudo().search(
                 [('code', '=', "env['ali.dindin.get.token'].get_token()")])
-            cron.sudo().unlink()
+            if cron:
+                cron.sudo().unlink()
